Fixed JumpTo(move)
Fixed JumpTo(move)
-After jumping backwards to a previouds board state,  the player's turn would still toggle normally instead of returning also to the turn of the player for that previous board state.

Separated all classes to its own file.
